+++ PowerEditor/src/TinyXml/tinyXmlA/tinyxmlparserA.cpp
@@ -395,7 +395,7 @@ void TiXmlDocumentA::StreamIn( TIXMLA_ISTREAM * in, TIXMLA_STRING * tag )
 
 	while ( in->good() )
 	{
-		int tagIndex = tag->length();
+		int tagIndex = static_cast<int>(tag->length());
 		while ( in->good() && in->peek() != '>' )
 		{
 			int c = in->get();
@@ -634,7 +634,7 @@ void TiXmlElementA::StreamIn (TIXMLA_ISTREAM * in, TIXMLA_STRING * tag)
 			// We should be at a "<", regardless.
 			if ( !in->good() ) return;
 			assert( in->peek() == '<' );
-			int tagIndex = tag->length();
+			int tagIndex = static_cast<int>(tag->length());
 
 			bool closingTag = false;
 			bool firstCharFound = false;
